# api/public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Benevolus",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/campaigns": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "List campaigns",
                "operationId": "listCampaigns",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by campaign status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "Filter by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaigns retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CampaignResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Create a new campaign",
                "operationId": "createCampaign",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCampaignRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Campaign created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CampaignResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/campaigns/{id}": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get campaign by ID",
                "operationId": "getCampaign",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CampaignResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Update campaign",
                "operationId": "updateCampaign",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCampaignRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Campaign updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Delete campaign",
                "operationId": "deleteCampaign",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Campaign deleted successfully"
                    }
                }
            }
        },
...
        "/api/campaigns/{campaign}/favorite": {
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Toggle campaign favorite",
                "operationId": "toggleCampaignFavorite",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite toggled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "favorited": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreCampaignRequest": {
                "required": [
                    "title",
                    "description",
                    "goalCents"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string"
                    },
                    "goalCents": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "expiresAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateCampaignRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string"
                    },
                    "goalCents": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "in_review",
                            "open",
                            "closed",
                            "rejected",
                            "finished"
                        ]
                    }
                },
                "type": "object"
            },
...
            "CampaignResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "goal": {
                        "type": "number"
                    },
                    "raised": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}